+++ lib/model/settings/defaults.dart
@@ -31,7 +31,7 @@ class SettingsDefaults {
   static const mustBackTwiceToExit = true;
   static const keepScreenOn = KeepScreenOn.viewerOnly;
   static const homePage = HomePageSetting.collection;
-  static const showBottomNavigationBar = true;
+  static const enableBottomNavigationBar = true;
   static const confirmDeleteForever = true;
   static const confirmMoveToBin = true;
   static const confirmMoveUndatedItems = true;