+++ lib/services/global_search.dart
@@ -7,11 +7,11 @@ import 'package:flutter/widgets.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
 class GlobalSearch {
-  static const platform = MethodChannel('deckers.thibault/aves/global_search');
+  static const _platform = MethodChannel('deckers.thibault/aves/global_search');
 
   static Future<void> registerCallback() async {
     try {
-      await platform.invokeMethod('registerCallback', <String, dynamic>{
+      await _platform.invokeMethod('registerCallback', <String, dynamic>{
         'callbackHandle': PluginUtilities.getCallbackHandle(_init)?.toRawHandle(),
       });
     } on PlatformException catch (e, stack) {