+++ lib/widgets/common/identity/aves_filter_chip.dart
@@ -78,7 +78,7 @@ class AvesFilterChip extends StatefulWidget {
   });
 
   static Future<void> showDefaultLongPressMenu(BuildContext context, CollectionFilter filter, Offset tapPosition) async {
-    if (context.read<ValueNotifier<AppMode>>().value == AppMode.main) {
+    if (context.read<ValueNotifier<AppMode>>().value.canNavigate) {
       final actions = [
         if (filter is AlbumFilter) ChipAction.goToAlbumPage,
         if ((filter is LocationFilter && filter.level == LocationLevel.country)) ChipAction.goToCountryPage,